Added webassembly for node.js
Also added a new subcategory 'node.js' under 'Projects'. While it is able to compile C code to WebAssembly modules, it isn't a compiler on its own, technically, because it uses LLVM under the hood. It's actually a toolkit, so it would also fit under 'Tools', but I picked 'Projects' instead because the 'node.js' category is a better fit there.