Added the "WebAssembly threads in Firefox" article.
Added the "WebAssembly threads in Firefox" article to the Tutorials section. This article shows you how to extend Python's Simple HTTP Server in order to return the necessary response headers to enable the SharedArrayBuffer in Firefox (these headers will become the norm in all browsers in the near future). The article then shows how to use WebAssembly threads to convert a user-supplied image to greyscale.